Better control of parameters
Read and use more parameters from the input polynomial file. Use rlim/alim to determine the batch factorization primorial. Use lbpr and lbpa to control maximum primes in a relation, and mfbr and mfba to control cofactorization limits. The code should work for a lbpr/a of 32 bits, but not for higher values.